requirements.txt
langchain
ipykernel
python-dotenv
langchain_community
pypdf
bs4
arxiv
pymupdf
wikipedia
jq
langchain-text-splitters
python-dotenv
langchain-openai
chromadb
langchain-community
sentence_transformers
langchain_huggingface
faiss-cpu
langchain_chroma
beautifulsoup4
streamlit
langchain-ollama
langchain_groq
langchain_core
fastapi
uvicorn
langserve
sse_starlette